+++ csharp/Profiler/SpeedScope/SpeedScopeEvent.cs
@@ -2,7 +2,7 @@
 
 public class SpeedScopeEvent
 {
-    public double At { get; internal set; }
+    public decimal At { get; internal set; }
     public int Frame { get; internal set; }
     public string Type { get; internal set; }
 }
